Remove tmux-resurrect
It mostly found it annoying that I had to wait a while when I first
launched tmux. Furthermore it would restore the last active window which
in most cases wasn't what I wanted.

It also stopped working when I moved to a new laptop and I don't want to
spend the energy to fix it.