Set BROWSER environment variable
Use a custom script to open Firefox Developer Edition for localhost.
This allows me to use FireFox Developer Edition for local development
but my default browser (Firefox) for everything else.

This is necessary because both create-react-app and GitHub CLI look for
the BROWSER environment variables and I want it to mean different things
for each tool.