Switch to gruvbox theme
Time for a change of pace. I was fine with the whitescale theme but also
began noticing that it didn't fulfill it's purpose anymore.

The main goal of switching to whitescale was to force myself to read the
code more. Over time I just found different signals than color to
quickly jump through a document and still didn't end up reading the code
properly.

After seeing https://zaiste.net/posts/shell-commands-rust/ I feel like
it's time for a change of pace :D