Automatically switch terminal to dark/light mode
When the system changes to dark or light mode also automatically change
kitty, tmux and vim to the new mode.

This was heavily inspired by
https://arslan.io/2021/02/15/automatic-dark-mode-for-terminal-applications/.

The main difference is that instead of listing all vim instances through
tmux this uses neovim's rpc implementation to change the background.

To debug the launch service `tail -f`:
/tmp/dark-mode-notify-stderr.log
/tmp/dark-mode-notify-stdout.log

Note that several things are hardcoded:

- The socket kitty listens on is set to `/tmp/kitty` in `bin/theme`
  and `config/kitty/macos-launch-services-cmdline`
- The socket neovim listens on is set to `/tmp/nvim` in `zhsrc` and
  `bin/neovim-background`
- The python binary used to call `bin/neovim-background` is set to
  `~/.asdf/shims/python` to ensure it's called with my preferred python
  version

Hardcoding these is unfortunately necessary because the `theme` script
is ultimately meant to be run by launchctl. Launchtl does not have
access to my user's environment variables. This means it wouldn't know
where to look for the kitty and neovim sockets and use the system
default python 2.7 which does not have pynvim installed.