Make vim depend on system dark mode
Even with automatically setting vim, tmux and kitty to dark/light mode
it was still possible to get vim in an incorrect state.

Specifically when the appearance was changed while a shell was open
$THEME would be set to the original value. Because vimrc reads $THEME
vim would open using the appearance when the shell was open.

In rough steps:

1. Assume $THEME is light
2. Open kitty
3. Change system appearance
5. Open vim

Vim would open with a light background because changing the system
appearance does not change the $THEME environment variable for shells
that are already open.