Strip new line from command output
Running defaults read returns the current interface style with a newline
at the end. Vim converts this newline into a null character (^@) which
makes the conditional fail.

Using printf in system strips the newline and ensures we get a string
with no special character.

See https://superuser.com/questions/935574/get-rid-of-null-character-in-vim-variable