Use a custom runner for Sketch's frontend codebase
This is a copy of vim-test's reactscripts test code where the test
executable is changed to `yarn test --watchAll=false` instead of
`react-scripts test`

This is very specific to the Sketch frontend codebase so I decided to
create a plugin. The plugin is used only when we're in a package named
`sketch-cloud-frontend`.