Initial NeoVim init with lazy.nvim
I figured if I go all new I might as well use the latest hip things.
Which means configuring NeoVim with Lua and using lazy.nvim as a package
manager.

I picked lazy.nvim over packer.nvim because people on the internet
seemed to like it more. I also enjoyed the clarity of it's README
compared to that of Packer.

Somewhat arbitrary.