Show hidden characters
I want to know if a file is indented using spaces or tabs so I make
those characters visible in vim. I also want to know if a line has
trailing space so I show trailing spaces _and_ the end of line character
to make it clearer how much trailing space there is.