Make git commit signing work on non-macOS systems
The gitconfig was still directly pointing to the macOS location of
op-ssh-sign.

I'm not sure how to make that config platform independent so I decided
to remove it and set SSH_AUTH_SOCK instead. This works _except_ that
Ubuntu/1Password insist on showing me a GUI dialog to authenticate.
Which doesn't work over SSH.