Configure 1Password to sign git commits
This was already configured and working on macOS and it has previously
worked on Ubuntu. It however stopped working and the only way I can get
it working again is by configuring the gpg["ssh"].program in gitconfig.

Unfortunately the location of op-ssh-sign differs per operating system.
This introduces a small helper executable that calls the correct
underlying executable depending on the system.

This also adds ~/.bin to $PATH so git can find the helper
executable.